% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_client.R
\docType{class}
\name{PredictionClient-class}
\alias{PredictionClient-class}
\alias{PredictionClient}
\title{Prediction Client Class}
\description{
Prediction client
}
\details{
Prediction Client to help capture prediction data
}
\section{Fields}{

\describe{
\item{\code{feature_names}}{array of feature names}

\item{\code{predict_names}}{array of prediction names}

\item{\code{metadata_names}}{array of metadata names}

\item{\code{instance_id}}{string containing a unique instance id}
}}


\examples{
library(DominoPredictionLogging)
prediction_client <- PredictionClient(
feature_names = c("sepal.length", "sepal.width", "petal.length"),
predict_names = c("variety")
)
prediction_client$record(
feature_values = c(sepal_length, sepal_width, petal_length),
predict_values = c(predicted_value),
event_id = event_id
)
}
